wip: fix Settings service
